requirements.txt
torch
torchvision
opencv-python
Pillow
playsound==1.2.2

